Fix distance float parsing error and implement dynamic class filters in spellbook

+++ app/src/main/java/com/example/dndcompanion/ui/viewmodel/SpellbookModels.kt
@@ -62,7 +62,7 @@ data class SpellDto(
 
 data class SpellRangeDto(
     val type: String,
-    val distance: Int
+    val distance: Double
 )
 
 data class SpellComponentsDto(
